Move MockYAML to conftest and patch classic_settings()
